% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bestan.R
\name{bestan}
\alias{bestan}
\title{BEST: Stan edition}
\usage{
bestan(y1, y2, iters = 110000)
}
\arguments{
\item{y1}{A vector of values for group 1}

\item{y2}{A vector of values for group 2}

\item{iters}{Number of MCMC iterations. Defaults to same as original BEST.}
}
\description{
Use bestan() to compare two groups.
}
\examples{
group_1 <- rnorm(20, 10, 5)
group_2 <- rnorm(20, 0, 2)
bestan(y1=group_1, y2=group_2)
}

